Retrofits after testing

+++ src/services/summarizer/pdf.py
@@ -8,53 +8,39 @@ from google.genai import types
 
 from services.setting.typex import ISettingService
 from services.logger.typex import ILoggerService
-from services.summarizer.typex import ISummarizerService, SummarizerResult, produce_summarization_prompt, process_response, create_fallback_structure
+from services.summarizer.typex import ISummarizerService, SummarizerResult, produce_title_prompt, produce_summarization_prompt, process_response, create_fallback_structure
 
 
 class PdfSummarizerService:
-    def __init__(self, setting: ISettingService, logger: ILoggerService):
+    def __init__(self, setting: ISettingService, logger: ILoggerService, text_summarizer_service: ISummarizerService):
         self._setting = setting
         self._logger = logger
+        self._text_summarizer_service = text_summarizer_service
 
     async def summarize(
             self,
             channel: str,
-            message: str,
             category: str,
-            language: str = None,
-            title: Optional[str] = None,
+            language: str,
+            message: str,
             metadata: Optional[Dict[str, Any]] = None
         ) -> Optional [SummarizerResult]:        
         """
-        Process an incoming message and create a note.
+        Process an incoming message and create a summary.
 
         Args:
             channel: The channel to which the summary belongs
-            message: The incoming message
             category: The category of the note
             language: The language of the message
-            specified_title: The title of the note (if specified)
-            specified_category: The category of the note (if specified)
+            message: The message content
             metadata: Optional dictionary containing additional metadata for the note
-
-        Returns:
-            SummarizerResult: A structured result containing the summary and metadata.
         """
+        # pdf only works with Gemini models
+        if self._setting.get_summarization_llm_provider() != "gemini":
+            self._logger.error(f"Only Gemini models are supported.")
+            return create_fallback_structure(message, language, channel, category)
+
         try:
-            # pdf only works with Gemini models
-            if self._setting.get_summarization_llm_provider() != "gemini":
-                self._logger.error(f"Gemini model is not available.")
-                return create_fallback_structure(message, language, title, category)
-
-            client = genai.Client()
-
-            skip_summarization = False
-            if category and category.lower() in ["sayings", "poetry"]:
-                skip_summarization = True
-
-            prompt = produce_summarization_prompt(category, language, message, skip_summarization)
-            self._logger.debug(f"Prompt: {prompt}")
-
             pdf_path = message.strip()
             self._logger.info(f"Analyzing file: {pdf_path}")
 
@@ -75,38 +61,44 @@ class PdfSummarizerService:
             shutil.copy2(pdf_path, temp_path)
 
             # Upload using the temp file (ASCII-safe path)
+            client = genai.Client()
             uploaded_file = client.files.upload(file=temp_path)
             self._logger.info(f"Upload complete. Remote File Name: {uploaded_file.name}")
 
-            try:
-                response = client.models.generate_content(
-                    model=self._setting.get_summarization_model(),
-                    contents=[uploaded_file, prompt]
-                )
+            metadata = {
+                "source": "pdf",
+                "reference": pdf_path
+            }
 
-                return process_response(message,
-                    response.text,
-                    channel,     
-                    language,
-                    category,
-                    title,
-                    {
-                        "source": "pdf",
-                        "reference": pdf_path
-                    })
+            prompt = produce_summarization_prompt(language)
+            self._logger.debug(f"PDF prompt: {prompt}")
 
-            except Exception as e:
-                print(f"An error occurred during generation: {e}")
+            response = client.models.generate_content(
+                model=self._setting.get_summarization_model(),
+                contents=[uploaded_file, prompt]
+            )
 
-            finally:
-                # Cleanup remote storage.
-                # Deleting the file from the API server after processing frees up your project storage quota.
-                client.files.delete(name=uploaded_file.name)
+            # produce consistent titles across all sources
+            title_prompt = produce_title_prompt(language, response.text)
+            self._logger.debug(f"PDF title prompt: {title_prompt}")
+            title_summary = await self._text_summarizer_service.summarize(channel, category, language, response.text)
+
+            return process_response(response.text,
+                title_summary.title,
+                language,
+                channel,     
+                category,
+                metadata)
 
         except Exception as e:
             self._logger.error(f"Error processing message with Gemini: {e}")
-            return create_fallback_structure(message, language, title, category)
+            return create_fallback_structure(message, language, channel, category)
         finally:
+            # Cleanup remote storage.
+            # Deleting the file from the API server after processing frees up your project storage quota.
+            if uploaded_file and client:
+                client.files.delete(name=uploaded_file.name)
+
             # Cleanup temporary local file
             if os.path.exists(temp_path):
                 os.unlink(temp_path)
